~Add incounters for child procces, to check exit status in pipeline for the right command;

+++ src/parser/type/heredoc.c
@@ -6,7 +6,7 @@
 /*   By: osivkov <osivkov@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/19 15:01:49 by osivkov           #+#    #+#             */
-/*   Updated: 2025/04/01 17:41:28 by osivkov          ###   ########.fr       */
+/*   Updated: 2025/04/02 15:26:07 by osivkov          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,69 +16,42 @@
 #include <readline/readline.h>
 #include <readline/history.h>
 
-static int	ft_strcmp(const char *s1, const char *s2)
+void	heredoc_child(t_heredoc_context *ctx, char *delimiter)
 {
-	while (*s1 && (*s1 == *s2))
-	{
-		s1++;
-		s2++;
-	}
-	return ((unsigned char)*s1 - (unsigned char)*s2);
+	if (!process_heredoc_lines(ctx, delimiter))
+		exit(1);
+	finalize_heredoc(ctx);
+	exit(0);
 }
 
-static void	cleanup_heredoc_error(t_heredoc_context *ctx)
+int	heredoc_parent(t_heredoc_context *ctx, pid_t pid)
 {
-	dup2(ctx->backup_fd, STDIN_FILENO);
-	close(ctx->backup_fd);
-	close(ctx->temp_fd);
-	unlink(".here_doc_tmp");
+	int		status;
+	int		fd;
+
 	signal(SIGINT, ctx->old_handler);
-}
-
-static int	process_heredoc_lines(t_heredoc_context *ctx, char *delimiter)
-{
-	char	*line;
-
-	while (1)
-	{
-		line = readline("heredoc> ");
-		if (!line)
-			return (cleanup_heredoc_error(ctx), 0);
-		if (g_exit == 130)
-		{
-			free(line);
-			cleanup_heredoc_error(ctx);
-			g_exit = 0;
-			return (0);
-		}
-		if (ft_strcmp(line, delimiter) == 0)
-		{
-			free(line);
-			break ;
-		}
-		write(ctx->temp_fd, line, ft_strlen(line));
-		write(ctx->temp_fd, "\n", 1);
-		free(line);
-	}
-	return (1);
-}
-
-static int	finalize_heredoc(t_heredoc_context *ctx)
-{
+	waitpid(pid, &status, 0);
 	dup2(ctx->backup_fd, STDIN_FILENO);
 	close(ctx->backup_fd);
 	close(ctx->temp_fd);
-	ctx->temp_fd = open(".here_doc_tmp", O_RDONLY);
-	if (ctx->temp_fd < 0)
-		perror("Error: unable to open temporary file for here_doc");
+	if (!WIFEXITED(status) || WEXITSTATUS(status) != 0)
+	{
+		unlink(".here_doc_tmp");
+		return (-1);
+	}
+	fd = open(".here_doc_tmp", O_RDONLY);
+	if (fd < 0)
+	{
+		unlink(".here_doc_tmp");
+		return (-1);
+	}
 	unlink(".here_doc_tmp");
-	g_exit = 0;
-	signal(SIGQUIT, SIG_DFL);
-	return (ctx->temp_fd);
+	return (fd);
 }
 
 int	handle_heredoc(char *delimiter)
 {
+	pid_t				pid;
 	t_heredoc_context	ctx;
 
 	ctx.backup_fd = dup(STDIN_FILENO);
@@ -86,11 +59,19 @@ int	handle_heredoc(char *delimiter)
 	ctx.temp_fd = open(".here_doc_tmp", O_WRONLY | O_CREAT | O_TRUNC, 0644);
 	if (ctx.temp_fd < 0)
 	{
-		perror("Error: Unable to create temporary file for here_doc");
+		perror("Error: Unable to create temporary file for heredoc");
 		signal(SIGINT, ctx.old_handler);
 		return (-1);
 	}
-	if (!process_heredoc_lines(&ctx, delimiter))
+	pid = fork();
+	if (pid < 0)
+	{
+		perror("Fork error");
 		return (-1);
-	return (finalize_heredoc(&ctx));
+	}
+	else if (pid == 0)
+		heredoc_child(&ctx, delimiter);
+	else
+		return (heredoc_parent(&ctx, pid));
+	return (-1);
 }
